+++ entity/manufacturer.go
@@ -1,6 +1,6 @@
 package entity
 
 type Manufacturer struct {
-	ID   uint16 `json:",string"`
+	ID   int64 `json:",string"`
 	Name string
 }
